pagecache_deallocate_node(): properly free allocated memory
This commit provides a proper implementation for
pagecache_deallocate_node(), so that no memory is leaked when a
page cache node is no longer used. This function can only be called
when no I/O operations are pending on any pages of the node; to
comply with this requirement, calls to pagecache_sync_node() have
been added to the code that handles deallocation of fsfile
structures and TFS log extensions. As a side effect, deallocation
of the storage space occupied by a file being deleted may now be
done asysnchronously, which required the unlink runtime test to be
amended in order to give time to the kernel to do this deallocation
before checking the available storage space.
In destroy_filesystem(), the call to pagecache_dealloc_volume() has
been moved to after the calls to deallocate_fsfile(), because page
cache nodes reference the containing page cache volume and thus the
volume can only be deallocated after all its nodes have been
deallocated.