mmap(): allow safe access to file-backed memory after file closure
Any file-backed memory mapping should be valid even after the file
descriptor has been closed (and the file possibly deleted). This
commit adds reference counting to ensure that a memory-mapped file
is not deallocated from the storage device; an fsfile field has
been added to the vmap structure, so that the file reference count
can be adjusted when the vmap is allocated and deallocated.
A new test case has been added to the mmap runtime tests to
exercise the code path that deallocates a file during unmapping.