Xen: add support for watching xenstore paths
If a given xenstore patch is watched, an asynchronous watch event
is delivered on the xenstore event channel when there are changes
at or below the path.
The code that makes synchronous requests to the xenstore is being
refactored to be able to handle asynchronous events being delivered
in the middle of a request.
This new feature will be used to support xen device hotplugging.