Xen: add support for device hotplugging
This commit adds support for detecting Xen devices added to the
xenstore after boot; to implement this, the "device" xenstore path
is watched so that any modifications to the list of devices in the
xenstore are notified asynchronously. The code that scans the list
of devices has been modified so that it can be called more than
once and updates the xen_info.device_tree tuple set after the first
call.
Any existing bindings between a device and a driver are recorded
via the sym(bound) tuple entry, so that a probe closure is not
called on a device that has already been bound to a driver.
Any devices that have been removed from the xenstore are removed
from the xen_info.device_tree tuple set in xen_probe_id_each(),
which checks the backend state and removes the device if the state
either cannot be retrieved or is `XenbusStateClosed`.