TFS: put all directory entries in internal file table
With this change, all directory entries (not just regular files)
are put in the file table. Regular files are distinguished from
non-regular files by the fact that the latter have a file table
entry value set to INVALID_ADDRESS (as opposed to an fsfile
pointer).
This will be needed in the next commit to be able to validate inode
numbers supplied as arguments to TFS functions.