TFS: add protection against concurrent execution
This will allow filesystem-related syscalls and I/O completions to
be run without the global kernel lock.
A call to thread_log() has been moved out of fill_stat() and placed
in the functions that call fill_stat() after the call to
filesystem_put_*(): this is necessary because printing trace
messages while a filesystem is locked can lead to a deadlock if the
syslog console driver is enabled and configured to output to a
file.