Unix signals: make code safe to run without the kernel lock
In an SMP system, without the kernel lock multiple cores can call
dequeue_signal() concurrently; since this function accesses the
process signal state, it can happen that a signal initially
selected among the pending signals cannot be dequeued because it
has been dequeued by another core; in this case, instead of
triggering an assertion failure, another attempt should be made at
finding a signal to be dequeued; if a signal is successfully
dequeued, the corresponding sigstate struct should be kept locked
until no longer accessed.
In deliver_signal(), the sigstate struct should be kept locked
during the entire sequence of instructions that access it; if the
sigstate is unlocked after determining that a signal should be
delivered but before setting the signal as pending, it could happen
for example that another core sets the same signal as pending,
causing the signal to be enqueued twice.
In sigstate_thread_restore(), the sigstate being restored should be
locked, because it could be the process-wide sigstate.

Use of the general heap is being removed in favor of the locked
heap.