Unix sockets: fix usage of transmission blockq
When a socket write operation is carried out, it should be tied to
the tx blockq of the destination socket, not the one of the sender,
so that if the destination queue becomes full and a write operation
blocks, the receiver can notify the sender via blockq_wake_one()
when the queue becomes writable again, without needing a reference
to the sender (for datagram sockets, there is no such reference).
In addition, for poll operations to be supported on datagram
sockets, the sender should add a notify handler to the receiver
notify set.
Whenever a reference to a peer socket is being used (such as during
a connect or a sendto), its reference count is now incremented,
regardless of the socket type, so that it is safe to access the
peer socket structure even after the peer socket has been closed
(and it is possible to detect if the peer socket has been closed by
checking the data field). Socket path resolution has been moved
from unixsock_write_bh() to unixsock_sendto(). These changes will
make it easier to implement multi-thread safety without the global
kernel lock.
Handling of connect() and poll() operations, including various
error conditions, has been revised to make it more consistent with
the Linux implementation, and a number of test cases have been
added to the runtime test.