revise timerqueue (formerly timerheap) interface and protect pqueue and protect pqueue access with spinlocks
This commit makes several changes to the interface used to register and
maintain kernel timers. For one, timer objects are no longer allocated on each
call to register_timer(), but are instead allocated by the caller (typically
embedded as part of an object that makes use of the timer) and passed to
register_timer() for insertion into the timerqueue. Secondly, timer handler
callbacks now include the expiry that was used on timer registration. This may
be useful for situations where a large number of timers may be served by a
single timer_handler callback, as the callback can use the (absolute) expiry
to associate the call with the corresponding event registered. Third, calls to
remove_timer() either atomically remove a timer from the timerqueue and return
true, or else return false if the timer was not registered or a one-shot timer
had already fired. Canceled timers result in a timer_handler callback with an
overrun field set to timer_disabled (-1ull). The forthgoing pattern is to only
assume that a timer is finally unregistered with the timer service when either
1) a one-shot timer has fired, or 2) a canceled one-shot or interval timer
resulted in a callback with timer_disabled. This avoids synchronization issues
that may occur between multiple execution paths (the code cancelling the timer
and the path resulting in the timer_handler invocation). Finally, operations
on the timerqueue are now protected with a spinlock. However, handler
callbacks are not covered by this spinlock, avoiding the possibility of a
deadlock should handlers in turn try to register or remove timers on the same
queue.

To facilitate the removal of timer entries, the pqueue_remove() method has
been added to the pqueue data structure, along with corresponding removal
checks in the pqueue unit test. Incidentally, a minor path fix has been
included for the gcov build for unit tests.