I/O uring: make code safe to run without the global kernel lock
The fdesc lock is being reused instead of having an
io_uring-specific lock; in addition, interrupts are not disabled
when taking the lock, because none of the io_uring functions can
run in interrupt context.
Timeout handling is being revised to take into account that it may
not be possible to remove a kernel timer sysnchronously in SMP
machines, and when a timeout handler is invoked is must be able
(unless called with timer_disabled) to access its bound io_uring
struct.