Volumes: fix handling of synchronous filesystem creation completion
If the completion passed to create_filesystem() is invoked
synchronously (e.g. if an error is encountered), volume_link() is
called with the storage lock taken, thus it cannot call
storage_check_if_ready() which needs to acquire the lock.
To fix this issue, a boolean `mounting` variable is being added to
the global storage struct; when this variable is true, it means
that the volume or the mount point list is being iterated and thus
storage_check_if_ready() should not be called. To prevent missing
calls to storage_check_if_ready() (e.g. if a filesystem creation
process is completed while a different volume is being added), this
function is also called by volume_add() and
storage_set_mountpoints().
For improved efficiency, storage_check_if_ready() checks whether
the mount_complete closure status handler is valid before iterating
the volume list.