net/direct.c: fix packet reception for TCP clients
When a TCP client connection request is accepted by the server, the
direct_conn structure being created should be added to the
connection list of the direct structure, otherwise the receive
service does not process any packets received on the connection.
This commit moves from direct_accept() to direct_conn_alloc() the
insertion of the direct_conn element in the connection list, so
that this step is done for both clients and servers.