net/direct.c: apply connection handler in receive service routine
In order to allow TCP clients or servers to send data to a peer
from within the connection handler (as is done e.g. in the radar
and cloud_init klibs), the connection handler is called by the
receive service routine (where the lwIP lock is not held), instead
of by direct_conn_alloc() (where the lwIP lock is held).