Klibs: fix handling of lwIP lock during DNS resolution
When a host name needs to be resolved to an IP address, the DNS
callback provided to lwIP is invoked with the lwIP lock held.
To avoid deadlock, the DNS callbacks in the klibs have been changed
so that either no network packets are sent directly from within the
callback (as in the radar and cloud_init klibs, where network
requests are scheduled to be retried after a delay), or network
packets are sent without taking the lwIP lock (as in the ntp klib).