fsfile_open_or_create(): fix handling of file paths
A file path supplied to fsfile_open_or_create() can consist of a
simple file name not prefixed by a directory path; in this case,
filesystem_mkdirpath() should not be called.
This commit fixes an issue where fsfile_open_or_create() called
with a simple fine name would create a directory with the supplied
name, instead of creating a file.