Unix sockets: fix behavior of connect() before server accept()
When a server socket is in listening state, client connections
should succeed without blocking (unless the server backlog is
full), even if accept() has not been called on the server socket.
This commit fixes unixsock_connect() and unixsock_accept() to
implement the above behavior.
In addition, the EPROTOTYPE error code is returned instead of
ECONNREFUSED when trying to connect a given socket to a socket of a
different type, to match the Linux kernel behavior.