poll(): properly initialize revents field for all pollfd structures
When poll() is invoked, the kernel should initialize to 0 the
revents field for all the structs supplied in the syscall, not only
for those to be ignored, otherwise there may be spurious values in
this field when the syscall returns.
This commit fixes the above issue and adds to the pipe runtime test
a test case that would have failed without this fix.
Closes #833.