remove timer service from runloop, instead enqueuing a timer service routine to bhqueue on timer interrupt; rename runloop_timers to kernel_timers accordingly
In the effort to eliminate the kernel lock, the runloop_timers (now renamed
kernel_timers) are no longer serviced on each pass through the runloop under
the kernel lock. Instead, timer interrupts now register a universal
"timer_interrupt_handler" thunk which, in turn, schedules a timer service
routine to run from off of the bhqueue. This should also help limit checking
of the timer queue (which is now more costly given that a spinlock is taken to
check it) to follow timer interrupt events. In a similar vein, timer
re-programming is explicitly marked to occur (along with the soonest expiry)
in register_timer() and timer_service(), avoiding the need to check the timer
queue in update_timer().