Futex: fix race between FUTEX_WAIT and FUTEX_WAKE
In absence of the global kernel lock, there is a potential race
between FUTEX_WAIT and FUTEX_WAKE, which can manifest as in the
following example:
- thread 1 calls FUTEX_WAIT, locks the futex, checks that *uaddr
== 0, futex_bh() unlocks the futex and returns
BLOCKQ_BLOCK_REQUIRED
- thread 2 sets *uaddr = 1, calls FUTEX_WAKE, locks the futex,
tries to wake up a blocked thread but does not find any in the
futex bq waiter list, unlocks the futex
- in thread 1, blockq_check_timeout() processes the
BLOCKQ_BLOCK_REQUIRED return value from futex_bh(), adds the thread
to the futex bq waiter list, and goes to sleep, thus missing the
FUTEX_WAKE event from thread 2

This commit fixes the above race by adding a `waiters` field to the
futex struct, which tracks the number of threads waiting on the
futex: when a FUTEX_WAKE operation is requested, if
blockq_wake_one() cannot find a thread to wake up but the `waiters`
value is non-zero, blockq_wake_one() is called again, on the
assumption that a thread is about to be placed in the blockq waiter
list.
Since futex_bh() needs to modify the `waiters` value, in order for
the correct futex struct to be used by futex_bh() after one or more
threads are requeued from one futex to another, a blockq action
handler parameter has been added to blockq_transfer_waiters(): this
handler is invoked for each action being moved from one blockq to
another; the futex code implements this action so that the futex_bh
closure member that references the futex is set to the destination
futex of the requeue operation.

In addition, the `assert(current->blocked_on)` instruction in
thread_sleep_interruptible() has been removed, because without the
global kernel lock it is possible for a thread wakeup to be
triggered before or during invocation of this function, for example
when FUTEX_WAIT and FUTEX_WAKE are done simultaneously by two
different CPUs.