epoll_blocked_release(): fix locking of the list of blocked threads
When an epoll_blocked structure is being removed from an epoll
instance, the lock protecting the list of blocked threads in the
epoll instance must be acquired by epoll_blocked_release() not only
if the function call comes from a timeout or an interrupting
signal, but also when the function is called directly by the top
half of a syscall.
This commit changes epoll_blocked_release() to take the blockq
flags as a parameter and determine when to acquire and release the
lock based on the flags value.