Klibs: syslog: send UDP messages from flush timer
This change removes sending UDP syslog messages directly from the
console write handler, and instead sends them via the flush timer,
which has been repurposed to serve buffered data for both file and
UDP output.
This allows the syslog klib to send UDP messages output by lwIP
callbacks that write to the console, e.g. when a network interface
is assigned an IP address. In such cases, the lwIP lock is already
held when the console write handler is invoked, thus lwIP functions
that need the lock to be acquired cannot be called.