x86_64 stack protector: switch from global variable to TLS
This commit adds a stack canary field to struct cpuinfo_machine and
changes compiler flags so that the stack protector reads the canary
value from a per-CPU memory pointer instead of a global variable.
The function that initializes the stack protector canary has been
changed to take a cpuinfo parameter and is now implemented in
architecture-specific code.
In multi-processor systems, using TLS for the stack protector
provides added security, because exposure of the canary value used
in one processor cannot compromise code executed in other
processors; in addition, this will make it possible to implement
further security measures such as having different canary values
for different kernel execution contexts, or changing canary values
on a periodic basis.
Closes #730.