Allow /proc/mounts to notify pollers of mount change
The /proc/mounts file has special behavior where a change to the file
from a mount or unmount causes poll to return POLLPRI and POLLERR. This
can be used to monitor the mount status of the system. This change
implements this feature by adding a generation number to track changes
to mounts and calling a list of registered handlers whenever a mount
changes. The /proc/mounts special file registers a handler that checks
to see if the generation number has changed and calls notify_dispatch
to send the event to waiters on the file. Special files structs are now
allocated differently so that they can wrap their own struct around
the file struct.