klibs: cloud_init: fix retry timer instantiation
In the cloud init klib, a download retry timer is specific to a
given resource to be downloaded, thus there can be more than one
timer active at any given time. For this reason, a single static
struct timer cannot be used; instead, a separate struct needs to be
created for each download retry.
This commit replaces the static struct timer with a struct embedded
in the timer handler closure.