klibs: implement dynamic linking to kernel
This commit implements dynamic linking of klibs to the kernel,
which uses the symbol definitions found in the kernel ELF file and
removes the need for klibs to look up or define symbols explicitly.
Symbols defined in a klib are added to the main symbol table when
the klib has been loaded and successfully initialized (so that they
can be looked up by any klib that needs them), and are removed from
the symbol table when a klib is unloaded.

In order to prevent the linker from discarding functions that are
defined in the kernel but called exclusively by klibs, the list of
dynamic relocation entries found in each klib is retrieved via
objdump and transformed (using sed scripts) into a set of EXTERN
linker commands, which are put in an auto-generated file that is
included by the kernel linker script.

The now() kernel function cannot be dynamically linked to, because
it's an inline function that references VDSO variables, which are
declared with hidden visibility; for this reason, klibs have to
call the kern_now() wrapper function instead of now().

Since loading a klib requires the kernel symbol table to be
populated, if the kernel is not booted by stage2 (e.g. with
Firecracker) klib initialization is now done after loading the
symbols from the kernel ELF file, thus the boot filesystem needs to
be read regardless of whether the klibs are in that filesystem.

For simplicity, the `klib_md` parameter has been removed from the
init_klib() function, because it can be retrieved from the `fs`
argument value.