eliminate thread sighandler_frame, using single frame for all thread execution; correct signal restarting; support and test for sigsetjmp / siglongjmp
The previous signal handling construction requires switching to an alternate
frame (the "sighandler_frame") when dispatching a signal handler. As such, any
signal handling pattern which bypasses rt_sigreturn (such as the use of
siglongjmp) will not work, as doing so would prevent signal state stored in
the thread struct from being reverted, and thread execution would continue on
the sighandler_frame rather than returning to the default_frame. For the same
reasons, nested signal handling was also not supported.

This behavior is corrected by moving all thread execution to a single
frame (which, incidentally, is also needed for pending context-switching work)
and making use of the sigcontext stored on the thread stack to save and
restore program state. Special state held for the duration of signal handling
has been eliminated from the thread struct, thus enabling the use of
siglongjmp (which bypasses rt_sigreturn).

An issue in syscall restarting has also been resolved. Previously,
rt_sigreturn would check if a siginfo contained the SA_RESTART flag, applying
a check for -ERESTARTSYS in the position of the thread frame where a syscall
return value would be stored. This was done unconditionally, whether a syscall
was interrupted or not. This could pose a problem if a thread was interrupted
while running and by chance contained the value -ERESTARTSYS in the register
used for syscall return values. The solution here is to have
blockq_wake_internal_locked() set the "interrupting_syscall" flag in the
thread struct, which in turn later triggers the call to
check_syscall_restart() from dispatch_signals().

A test for sigsetjmp/siglongjmp use in signal handling has been added to the
signal runtime test.