CircleCI: upload build artifacts to Google Cloud bucket
This change adds a few steps to the "build-pc" CircleCI job, to
upload build artifacts to Google Cloud after a successful run of
the test suite. These steps are only executed if the
GCLOUD_SERVICE_KEY environment variable is set, so that the job
will not fail when triggered by pull requests from forked
repositories.
The mkfs and dump tools are not being included in the uploaded
artifacts, because Ops provides similar functionalities to create
and read image files; this means that the uploaded tarballs for
Linux and Mac are identical (we still need separate tarballs
because Ops uses the operating system name when downloading Nanos
releases from the cloud bucket).
The uploaded tarballs have the "Custom-Time" metadata set, so that
it's possible to delete them automatically from the bucket by
setting a "DaysSinceCustomTime" condition in the object lifecycle
management settings.