Runtime: implement generic ring buffer functionality
The implementation uses the same struct buffer that is used for
non-circular buffers, except that the start and end fields are
allowed to go past the buffer length. To retrieve the pointer to
the start of current buffer contents, the start field is masked
based on the total buffer length value (which is always a power of
2); the same applies to the end field.
A relevant set of tests has been added to the buffer unit tests.
This functionality will be used to implement the Linux inotify
syscalls.