Runtime tuples: add function to retrieve a symbol from a value
The new tuple_get_symbol() function iterates over the entries of
a tuple until it finds the supplied value, and then returns the
symbol associated to that value. Note: the return value is not
deterministic if the tuple contains more than one entry with the
supplied value.
A relevant set of tests has been added to the tuple unit tests.
This functionality will be used to implement the Linux inotify
syscalls.