getdents/getdetn64: remove duplicated code
Most of the code implementing the getdents and getdents64 syscalls
is identical, with the only difference being the userspace struct
(linux_dirent vs linux_dirent64) where directory entries are
placed. This commit refactors the code by defining shared internal
functions that handle both syscalls and take a boolean argument
indicating which of the two syscalls is being serviced.