TFS: refactor setting of directory modification time metadata
This commit adds setting of the modification time metadata in the
parent tuple in the fs_set_dir_entry() function. This allows
removing several calls to filesystem_update_mtime() in various
places that call fs_set_dir_entry(), and fixes a missing update of
the parent directory modification time in a few other places.