TFS: add reference to parent tuple to all filesystem entries
This commit adds to each filesystem tuple an entry (associated to
the ".." symbol) that references the parent tuple. This will be
needed for implementing the inotify functionality, in order to be
able to send notifications to the parent directory when an event
happens to a given filesystem entry.
This removes the need for the ".." entry in the children tuple of
directories; for uniformity, the "." entry is also being removed,
and the lookup() function is being amended to properly handle the
"." and ".." symbols.
Due to this change, before removing any filesystem entry (not just
directories) the reference to the parent tuple needs to be dropped;
for this purpose, the destruct_dir_entry() helper function has been
created.
In the getdents and getdents64 syscall implementation, the
getdents_internal() function has been amended in order to properly
list the "." and ".." directory entries.