Improve ACPI support with acpica
This commit adds full support for ACPI table parsing and ACPI
object evaluation and method execution by importing the acpica
library.
In order for the ACPI library to build correctly, a patch is
applied to it after cloning its git repository; to do this, the
`vendor` command in rules.mk has been enhanced so that it looks for
a patch file and if found applies it to the vendor code.
The glue code between acpica and the kernel is split in an
architecture-independent part (src/drivers/acpica.c) and an
architecture-specific part (implemented for x86_64 only at the
moment).
The x86_64-specific PIIX4 driver that was used to handle the
shutdown event and to power down the system has been replaced with
generic ACPI code. This brings support for the shutdown event and
for powering down the system to platforms that implement ACPI but
don't emulate a PIIX4 controller: for example, it is now possible
to power down gracefully a Qemu instance by enabling the Qemu
monitor and sending a `system_powerdown` command on the monitor
interface, and similarly a VirtualBox instance reacts to a shutdown
event and is able to power down itself.