PCI: add hotplug support
This commit adds support for PCI device hotplug, implemented via
ACPI control methods.
Since a hotplugged PCI device cannot be configured by the BIOS, the
kernel needs to be able to configure each PCI BAR needed for device
operation with an appropriate memory address range; this has been
implemented by retrieving from the ACPI tables the set of memory
address ranges assigned to each PCI host controller and using it to
initialize I/O memory heaps from which address ranges are allocated
when a PCI BAR needs to be configured.
When the hypervisor sends a request to detach a PCI device, an ACPI
GPE (General Purpose Event) is generated, which triggers an ACPI
interrupt; the interrupt handler in turn triggers the execution of
the notification handlers registered for the PCI device; the
handler implemented in the kernel then calls pci_remove_device();
the PCI driver interface has been enhanced with a `pci_remove`
closure, which is applied by pci_remove_device().
When a new PCI device is attached by the hypervisor, a similar
chain of events takes place, but in this case the notification
handler calls pci_probe_device(), which will then apply the driver
probe closure.
The NVMe controller driver implements the PCI remove closure and
thus supports volume detach operations executed on a running
instance. A call to pci_enable_io_and_memory() has been added in
the probe function because it is needed in order to enable I/O on
hotplugged NVMe controllers.
In kernel_runtime_init(), the call to init_interrupts() has been
moved so that it is made before the PCI bus is probed; this is
necessary because in the PC platform code init_interrupts() calls
init_acpi_tables(), which initializes the I/O memory heaps to be
assigned to the PCI host controllers. When a PCI device is probed,
for each enabled BAR with a memory address range, the address range
needs to be marked as allocated in the I/O memory heap of the host
controller.
A spinlock has been added in the generic PCI code to safely handle
concurrent calls to pci_probe_device() and pci_remove_device(),
including the case where multiple such calls are made for a given
device.