Kernel heaps: decrease utilization of physical memory
Instead of having a fixed-size bootstrap heap, which consumes
unnecessary memory in low-memory machines and does not allow
allocating all the available memory in machines with high amounts
of memory, this commit makes the bootstrap heap size dependent on
the amount of physical memory available at boot, and places the
base of the bootstrap heap at KMEM_BASE.
In addition, the sequence of initialization of the various kernel
heaps has been changed so that the bootstrap heap is only used as a
meta heap for the heaps that need to be created before the general
heap is available.
In order to decrease memory utilization, the locked heap is now a
locking heap wrapper around the general heap, instead of using its
own mcache.
The init_kernel_heaps() function has been moved from
platform-specific code to kernel/init.c, to remove duplicated code.