Multi-cache page size: add "low-memory" configuration
This commit changes the initialization of various multi-cache heaps
so that in machines with low amounts of physical memory a lower
cache page size is used instead of 2 MB. This allows the kernel to
run in low-memory machines, while not affecting performance in
high-memory machines.
The build_assert() statement in allocate_tagged_region() has been
removed because it no longer applies now that it is possible to
allocate in a multi-cache memory larger than the maximum cache page
size.