HTTP parser: add deallocation of parsed data and parser struct
This commit implements in the HTTP parser deallocation of parsed
data after applying the value handler. In addition, when the
http_recv closure is applied with a null buffer, the parser
deallocates itself and the value handler closure. This will allow
easy implementation of a web server that does not leak memory when
handling client requests.
Since deallocation of parsed data is done via a call to
destruct_tuple(), there must not be multiple tuple entries with the
same buffer value, otherwise this buffer would be deallocated
multiple times; for this reason, the each_http_request closure is
being modified to use a separate buffer to encode the relative URI,
instead of using the same buffer that contains the absolute URI.
Deallocation of parsed data is being removed from the klib
functions that use the HTTP parser.