unix_process/socket_user.c: define conn_handler struct
This commit changes the implementation of a network connection
input and output handlers so that all the data necessary to handle
a connection (including the input and output closures) is stored in
a single struct, which is deallocated when the connection is
closed. This fixes a memory leak caused by the input and output
closures not being deallocated.