Klibs: fix make dependencies to generate symbol list
Updating the symbol list in klib-syms.lds after a given klib binary
is built does not work well when make is run with multiple parallel
jobs, because different klib binaries can be built in parallel and
thus the symbol list file can be updated concurrently by multiple
jobs, which can lead to inconsistencies in the final contents of
the file and ultimately cause klib loading to fail due to missing
kernel symbols.
This commit changes the make target for generating the symbol list
so that this target has the klib binaries as dependencies, and thus
is executed only after all binaries have been built. In addition,
the following improvements have been made:
- the symbol list is cleared before being re-built: this eliminates
symbols that were used in previous versions of klibs but are no
longer used
- duplicated entries in the symbol list are removed only once,
instead of after building each klib binary
- similarly, the backup file is deleted only once, when the final
symbol list is ready