Global/static variables: add const modifier where appropriate
Adding the const modifier instructs the linker to put those
variables in the rodata section, which can be mapped as read-only
in the MMU and thus protected against unwanted modifications.

The `-n` linker flag, which turns off page alignment of ELF
sections, may cause issues because the boot code that maps the
kernel ELF binary expects the file offset of all sections to have
the same page offset as the base virtual address, and the `-n` flag
may cause this requirement to not be satisfied; thus, this flag is
being removed. In order to prevent the linker from aligning ELF
sections to 2MB boundaries (which would considerably increase the
kernel binary file size), the `-z max-page-size=4096` linker flag
is being added.