mutex implementation
Building off of the context switching work, this change implements a two-phase
mutex lock which will first spin for some number of iterations (specified on
mutex allocation) before finally suspending the running context if the mutex
could not be acquired. Suspended contexts are queued and processed in a FIFO
fashion, with the unlock function dequeueing and directly scheduling any
context waiting on the mutex. This is useful for preventing cores from being
excessively tied up with spinning while there is resource contention. The spin
factor argument given on instantiation may be adjusted according to the needs
of the workload; a value of 0 indicates no spinning is desired - in which case
a context will suspend whenever the mutex cannot be immediately acquired - but
may also be made large to prioritize spinning, leaving suspension as a
fallback option that frees up cpu cycles for other contexts to run.

The new context_suspend() function suspends a context by saving the running
processor state to its frame and then adjusting it to return to the caller on
frame return.

As a first client, the lwIP lock has been converted from a spinlock to a
mutex. It is instantiated with a large spin factor in order to maintain
existing performance levels observed with our webg test. As with spinlocks,
the factor of bus contention when considering performance increases with the
number of cores contending for a mutex. This could potentially be ameliorated
in the future by using something like the MCS lock in place of the
compare-and-swap loop for the first phase. Using MCS would also insure that
waiters are processed in a FIFO order for both phases of lock acquisition, not
just the second one.