net connection_handler: use input_buffer_handler for incoming data
The closure that handles incoming data on a given network
connection may close the connection (by invoking the output buffer
handler with a null buffer argument); if this happens, the
direct_conn struct is deallocated, thus it should not be accessed
anymore.
This change replaces the buffer_handler closure type for handling
incoming data with a new input_buffer_handler type, whose return
value is a boolean variable that indicates whether the connection
has been closed. The code in net/direct.c uses this return value to
ensure that the direct_conn struct is not accessed after closing a
connection. In addition, since in case of client connections the
associated direct struct is also deallocated, the
direct_conn_closed() function has been modified to return a boolean
value that indicates whether the direct struct has been
deallocated.
This change fixes a crash occurring on Microsoft Azure instances
when the cloud_init klib reports the instance status to the wire
server.