use lock-free list to implement queue of waiting contexts
Using a fixed-size queue of waiters and spinning on failing enqueue operations
ultimately defeats the mutex property that a context awaiting a mutex lock
cannot tie up a CPU indefinitely. To remedy this, the queue has been
substituted with a singly-linked list of waiting contexts. The lock-free list
follows the one used in the MCS lock, albeit for enqueuing waiters and not
spinning.