pending_fault_complete(): protect service of pending_fault dependents with p->faulting_lock
This fixes a race whereby a thread could be added to a pending_fault's
dependents vector during or after service of the vector in
pending_fault_complete(). Such an occurance could result in a thread hanging
indefinitely on a page fault. This moves the taking of p->faulting_lock in
pending_fault_complete() up prior to servicing the dependents vector.