remove utime_updated() and stime_updated()
The utime_updated() and stime_updated() functions may access the "start_time"
field of a thread or a thread's syscall_context, respectively. This is
potentially unsafe as these fields, as well as the thread's syscall field, may
be updated without holding the thread lock. These functions are also
unnecessary because a thread's execution times are accumulated on each context
switch, providing updates at a finer granularity (e.g. with each interrupt,
syscall entry, blocking or suspending of a syscall, etc.).